backend/requirements-lightweight.txt
# Lightweight requirements for PyInstaller backend
# Excludes ML dependencies (PyTorch, OpenSoundscape, etc.)
# Only includes essential libraries for data processing, audio handling, and basic utilities.

# Build tools
pyinstaller>=6.14.2

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Audio processing (without heavy ML deps)
librosa>=0.9.0
soundfile>=0.10.0
resampy>=0.3.0

# Image processing
Pillow>=8.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Data formats
h5py>=3.0.0
pyyaml>=6.0

# System utilities
psutil>=5.8.0
tqdm>=4.60.0
pathlib2>=2.3.0

# HTTP server
aiohttp>=3.8.0
aiohttp-cors>=0.7.0

# Configuration and validation
pydantic>=1.8.0

# Standard library enhancements
click>=8.0.0
python-dateutil>=2.8.0